# api/swagger.yaml
openapi: 3.0.1
info:
  title: Setu Addressing API - draft
  description: |
    1. Create payment address such as UPI VPA
    2. Set settlement rules
    3. Pause address
    4. Delete address
    5. Edit settlement rule
  termsOfService: http://seut.co
  contact:
    email: ranjith@setu.co
  version: 1.0.0
...
servers:
- url: https://sandbox.setu.co/api
- url: http://sandbox.setu.co/api
...
paths:
...
  /routes/addresses/{addressID}/transactions:
    get:
      tags:
      - Transactions
      description: List transactions
      operationId: listTransactions
      parameters:
      - name: X-PRODUCT-INSTANCE-ID
        in: header
        required: true
        schema:
          type: string
      - name: addressID
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      transactions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
                    example:
                      transactions:
                      - amountPaid:
                          value: 6
                          currencyCode: INR
                        platformTransactionRefID: platformTransactionRefID
                        uniquePaymentRefID: uniquePaymentRefID
                        timestamp: timestamp
                      - amountPaid:
                          value: 6
                          currencyCode: INR
                        platformTransactionRefID: platformTransactionRefID
                        uniquePaymentRefID: uniquePaymentRefID
                        timestamp: timestamp
                example:
                  data:
                    transactions:
                    - amountPaid:
                        value: 6
                        currencyCode: INR
                      platformTransactionRefID: platformTransactionRefID
                      uniquePaymentRefID: uniquePaymentRefID
                      timestamp: timestamp
                    - amountPaid:
                        value: 6
                        currencyCode: INR
                      platformTransactionRefID: platformTransactionRefID
                      uniquePaymentRefID: uniquePaymentRefID
                      timestamp: timestamp
                  success: true
                  status: 0
                x-content-type: application/json
      security:
      - Bearer: []
      x-swagger-router-controller: Transactions
  /routes/addresses/{addressID}/transactions/{transactionID}:
    get:
      tags:
      - Transactions
      description: Get beneficiary by ID
      operationId: getTransactionById
      parameters:
      - name: transactionID
        in: path
        required: true
        schema:
          type: string
      - name: addressID
        in: path
        required: true
        schema:
          type: string
      - name: X-PRODUCT-INSTANCE-ID
        in: header
        required: true
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      transaction:
                        $ref: '#/components/schemas/Transaction'
                    example:
                      transaction:
                        amountPaid:
                          value: 6
                          currencyCode: INR
                        platformTransactionRefID: platformTransactionRefID
                        uniquePaymentRefID: uniquePaymentRefID
                        timestamp: timestamp
                example:
                  data:
                    transaction:
                      amountPaid:
                        value: 6
                        currencyCode: INR
                      platformTransactionRefID: platformTransactionRefID
                      uniquePaymentRefID: uniquePaymentRefID
                      timestamp: timestamp
                  success: true
                  status: 0
                x-content-type: application/json
      security:
      - Bearer: []
      x-swagger-router-controller: Transactions
  /routes/addresses/{addressID}/refunds:
    get:
      tags:
      - Refunds
      description: List Refund requests
      operationId: listRefunds
      parameters:
      - name: X-PRODUCT-INSTANCE-ID
        in: header
        required: true
        schema:
          type: string
      - name: addressID
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      refunds:
                        type: array
                        items:
                          $ref: '#/components/schemas/Refund'
                    example:
                      refunds:
                      - ""
                      - ""
                example:
                  data:
                    refunds:
                    - ""
                    - ""
                  success: true
                  status: 0
                x-content-type: application/json
      security:
      - Bearer: []
      x-swagger-router-controller: Refunds
    post:
      tags:
      - Refunds
      description: create a refund request
      operationId: addRefund
      parameters:
      - name: X-PRODUCT-INSTANCE-ID
        in: header
        required: true
        schema:
          type: string
      - name: addressID
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseRefund'
        required: false
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      refund:
                        $ref: '#/components/schemas/Refund'
                    example:
                      refund: ""
                example:
                  data:
                    refund: ""
                  success: true
                  status: 0
                x-content-type: application/json
      security:
      - Bearer: []
      x-swagger-router-controller: Refunds
  /routes/addresses/{addressID}/refunds/{refundID}:
    get:
      tags:
      - Refunds
      description: Get refund by ID
      operationId: getRefundById
      parameters:
      - name: refundID
        in: path
        required: true
        schema:
          type: string
      - name: addressID
        in: path
        required: true
        schema:
          type: string
      - name: X-PRODUCT-INSTANCE-ID
        in: header
        required: true
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      refund:
                        $ref: '#/components/schemas/Refund'
                    example:
                      refund: ""
                example:
                  data:
                    refund: ""
                  success: true
                  status: 0
                x-content-type: application/json
      security:
      - Bearer: []
      x-swagger-router-controller: Refunds
components:
  schemas:
...
    Amount:
      type: object
      properties:
        value:
          type: integer
        currencyCode:
          type: string
          enum:
          - INR
      example:
        value: 6
        currencyCode: INR
...
    BaseRefund:
      type: object
      properties:
        platformTransactionRefID:
          type: string
    RefundRejectReason:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
    Refund:
      allOf:
      - type: object
        properties:
          id:
            type: string
          status:
            type: string
            enum:
            - PENDING
            - SUCCESS
            - REJECTED
          rejectReason:
            $ref: '#/components/schemas/RefundRejectReason'
      - $ref: '#/components/schemas/BaseRefund'
    Transaction:
      type: object
      properties:
        platformTransactionRefID:
          type: string
        uniquePaymentRefID:
          type: string
        timestamp:
          type: string
        amountPaid:
          $ref: '#/components/schemas/Amount'
      example:
        amountPaid:
          value: 6
          currencyCode: INR
        platformTransactionRefID: platformTransactionRefID
        uniquePaymentRefID: uniquePaymentRefID
        timestamp: timestamp
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header